requirements.txt
Faker
matplotlib
pandas
Pillow
PySide6_Addons
PySide6_Essentials
bcrypt
reportlab
